requirements.txt
Django==5.0.6
djangorestframework==3.15.2
channels==4.1.0
daphne==4.1.2
channels-redis==4.2.0
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.7
python-dotenv==1.0.1
aiogram==3.5.0
django-environ==0.11.2
requests==2.32.3
whitenoise==6.7.0
